Further cleanup of casting issues.

+++ src/mlpack/tests/det_test.cpp
@@ -65,9 +65,7 @@ BOOST_AUTO_TEST_CASE(TestComputeNodeError)
   *min_vals << 3 << 0 << 1;
 
   DTree<>* testDTree = new DTree<>(max_vals, min_vals, 5);
-  double true_node_error = -1.0 * exp(-(double) log((double) 4.0)
-					   - (double) log((double) 7.0)
-					   - (double) log((double) 7.0));
+  double true_node_error = -1.0 * exp(-log(4.0) - log(7.0) - log(7.0));
 
   BOOST_REQUIRE_CLOSE(testDTree->error_, true_node_error, 1e-10);
 
@@ -75,10 +73,8 @@ BOOST_AUTO_TEST_CASE(TestComputeNodeError)
   testDTree->end_ = 5;
 
   double node_error = -std::exp(testDTree->LogNegativeError(5));
-  true_node_error = -1.0 * exp(2 * log((double) 2 / (double) 5)
-			       -(double) log((double) 4.0)
-			       - (double) log((double) 7.0)
-			       - (double) log((double) 7.0));
+  true_node_error = -1.0 * exp(2 * log(2.0 / 5.0) - log(4.0) - log(7.0) -
+      log(7.0));
   BOOST_REQUIRE_CLOSE(node_error, true_node_error, 1e-10);
 
   delete testDTree;
